Fold three over-extracted helper files back into their callers
- Inline parseSongTimeSignature into clip-result-helpers (was a 23-line
  one-function file in shared/)
- Rename duplicate-misc-helpers to duplicate-focus-helpers and drop the
  stale name-utils re-exports; duplicate.ts imports name utils directly
- Merge handleQuantization + QUANTIZE_GRID into update-clip-notes-helpers
  alongside the other MIDI note operations
